+++ src/ksource/utils.h
@@ -20,8 +20,8 @@ double rand_norm();
 double *traslv(double *vect, const double *trasl, int inverse);
 double *rotv(double *vect, const double *rotvec, int inverse);
 
-long pt2pdg(char pt);
-char pdg2pt(long pdgcode);
+int pt2pdg(char pt);
+char pdg2pt(int pdgcode);
 
 double interp(double x, const double *xs, const double *ys, int N);
